+++ Runtime/XRManagementAnalytics.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.CompilerServices;
+using System.Diagnostics.CodeAnalysis;
 
 #if UNITY_EDITOR
 using UnityEditor;
@@ -28,6 +29,9 @@ namespace UnityEngine.XR.Management
 
         [Serializable]
         private struct BuildEvent
+#if UNITY_2023_2_OR_NEWER && ENABLE_CLOUD_SERVICES_ANALYTICS && UNITY_ANALYTICS
+            : IAnalytic.IData
+#endif //UNITY_2023_2_OR_NEWER && ENABLE_CLOUD_SERVICES_ANALYTICS && UNITY_ANALYTICS
         {
             public string buildGuid;
             public string buildTarget;
@@ -35,10 +39,32 @@ namespace UnityEngine.XR.Management
             public string[] assigned_loaders;
         }
 
+#if UNITY_2023_2_OR_NEWER && ENABLE_CLOUD_SERVICES_ANALYTICS && UNITY_ANALYTICS
+        [AnalyticInfo(eventName: kEventBuild, vendorKey: kVendorKey, maxEventsPerHour: kMaxEventsPerHour, maxNumberOfElements: kMaxNumberOfElements)]
+        private class XrInitializeAnalytic : IAnalytic
+        {
+            private BuildEvent? data = null;
+
+            public XrInitializeAnalytic(BuildEvent data)
+            {
+                this.data = data;
+            }
+
+            public bool TryGatherData(out IAnalytic.IData data, [NotNullWhen(false)] out Exception error)
+            {
+                error = null;
+                data = this.data;
+                return data != null;
+            }
+        }
+#endif //UNITY_2023_2_OR_NEWER && ENABLE_CLOUD_SERVICES_ANALYTICS && UNITY_ANALYTICS
+
         private static bool Initialize()
         {
 #if ENABLE_TEST_SUPPORT || !ENABLE_CLOUD_SERVICES_ANALYTICS || !UNITY_ANALYTICS
             return false;
+#elif UNITY_EDITOR && UNITY_2023_2_OR_NEWER
+            return EditorAnalytics.enabled;
 #else
 
 #if UNITY_EDITOR
@@ -78,9 +104,11 @@ namespace UnityEngine.XR.Management
                 buildTargetGroup = buildTargetGroup.ToString(),
                 assigned_loaders = loaderTypeNames.ToArray(),
             };
-
+#if UNITY_2023_2_OR_NEWER
+            EditorAnalytics.SendAnalytic(new XrInitializeAnalytic(data));
+#else
             EditorAnalytics.SendEventWithLimit(kEventBuild, data);
-
+#endif //UNITY_2023_2_OR_NEWER
 #endif //UNITY_ANALYTICS && ENABLE_CLOUD_SERVICES_ANALYTICS
 
         }