+++ Tests/Editor/AndroidManifestTests.cs
@@ -12,6 +12,12 @@ using UnityEngine.XR.Management;
 
 public class AndroidManifestTests
 {
+    private const string k_androidXmlNamespace = "http://schemas.android.com/apk/res/android";
+    private const string k_unityActivityName = "com.unity3d.player.UnityPlayerActivity";
+    private const string k_unityGameActivityName = "com.unity3d.player.UnityPlayerGameActivity";
+    private readonly List<string> k_activityPath = new List<string> { "manifest", "application", "activity" };
+    private readonly List<string> k_categoryPath = new List<string> { "manifest", "application", "activity", "intent-filter", "category" };
+
     private string tempProjectPath;
     private string xrManifestTemplateFilePath;
     private string xrLibraryManifestFilePath;
@@ -170,7 +176,7 @@ public class AndroidManifestTests
         var processor = CreateProcessor();
 
         // Initialize data
-        var overrideElementPath = new List<string> { "manifest", "application", "activity" };
+        var overrideElementPath = new List<string> { "manifest", "application" };
         var overrideElement1Attributes = new Dictionary<string, string>();
         var overrideElement2Attributes = new Dictionary<string, string>();
         var providers = new List<IAndroidManifestRequirementProvider>()
@@ -226,7 +232,7 @@ public class AndroidManifestTests
         var processor = CreateProcessor();
 
         // Initialize data
-        var overrideElementPath = new List<string> { "manifest", "application", "activity" };
+        var overrideElementPath = new List<string> { "manifest", "test-tag" };
         var existingElementAttributes = new Dictionary<string, string>()
         {
             { "name", "com.test.app" }
@@ -279,6 +285,141 @@ public class AndroidManifestTests
             $"Attribute count {attributeList.Count} in element doesn't match expected {expectedElementAttrributes.Count}");
 
         AssertAttributesAreEqual(nodes[0].Name, expectedElementAttrributes, attributeList);
+    }
+
+    [Test]
+    public void AndroidManifestProcessor_UpdateExistingActivityElementWithOverridenElement()
+    {
+        // Use the Assert class to test conditions
+        var processor = CreateProcessor();
+
+        // Initialize data
+        var existingElementAttributes = new Dictionary<string, string>()
+        {
+            { "name", k_unityActivityName }
+        };
+        var overrideElementAttributes = new Dictionary<string, string>()
+        {
+            { "isGame", "true" },
+            { "testOnly", "true" },
+        };
+        var providers = new List<IAndroidManifestRequirementProvider>()
+        {
+            new MockManifestRequirementProvider(new ManifestRequirement
+            {
+                SupportedXRLoaders = new HashSet<Type>
+                {
+                    supportedLoaderType
+                },
+                OverrideElements = new List<ManifestElement>()
+                {
+                    new ManifestElement()
+                    {
+                        ElementPath = k_activityPath,
+                        Attributes = overrideElementAttributes
+                    }
+                }
+            })
+        };
+
+        // Execute
+        processor.ProcessManifestRequirements(providers);
+
+        // Validate
+        var updatedLibraryManifest = GetXrLibraryManifest();
+        var nodes = updatedLibraryManifest.SelectNodes(string.Join("/", k_activityPath));
+        Assert.AreEqual(
+            1,
+            nodes.Count,
+            "Additional elements exist in the Manifest when expecting 1");
+
+        var attributeList = nodes[0].Attributes;
+        var expectedElementAttrributes = MergeDictionaries(existingElementAttributes, overrideElementAttributes);
+        Assert.AreEqual(
+            expectedElementAttrributes.Count,
+            attributeList.Count,
+            $"Attribute count {attributeList.Count} in element doesn't match expected {expectedElementAttrributes.Count}");
+
+        AssertAttributesAreEqual(nodes[0].Name, expectedElementAttrributes, attributeList);
+    }
+    
+    [Test]
+    public void AndroidManifestProcessor_UpdateAllActivityElementWithOverridenElement()
+    {
+        IgnoreIfGameActivityIsNotSupported();
+
+        // Use the Assert class to test conditions
+        var processor = CreateProcessor();
+        processor.UseActivityAppEntry = true;
+        processor.UseGameActivityAppEntry = true;
+
+        // Initialize data
+        var existingElementAttributes = new Dictionary<string, string>()
+        {
+            { "name", k_unityActivityName }
+        };
+        var overrideElementAttributes = new Dictionary<string, string>()
+        {
+            { "isGame", "true" },
+            { "testOnly", "true" },
+        };
+        var providers = new List<IAndroidManifestRequirementProvider>()
+        {
+            new MockManifestRequirementProvider(new ManifestRequirement
+            {
+                SupportedXRLoaders = new HashSet<Type>
+                {
+                    supportedLoaderType
+                },
+                OverrideElements = new List<ManifestElement>()
+                {
+                    new ManifestElement()
+                    {
+                        ElementPath = k_activityPath,
+                        Attributes = overrideElementAttributes
+                    }
+                }
+            })
+        };
+
+        // Execute
+        processor.ProcessManifestRequirements(providers);
+
+        // Validate
+        var updatedLibraryManifest = GetXrLibraryManifest();
+        var nodes = updatedLibraryManifest.SelectNodes(string.Join("/", k_activityPath));
+        Assert.AreEqual(
+            2,
+            nodes.Count,
+            "Additional elements exist in the Manifest when expecting 2");
+
+        foreach (XmlNode node in nodes)
+        {
+            var attributeList = node.Attributes;
+            var expectedElementAttrributes = MergeDictionaries(existingElementAttributes, overrideElementAttributes);
+            Assert.AreEqual(
+                expectedElementAttrributes.Count,
+                attributeList.Count,
+                $"Attribute count {attributeList.Count} in element doesn't match expected {expectedElementAttrributes.Count}");
+
+            foreach (XmlAttribute attrib in attributeList)
+            {
+                var attributeName = attrib.Name.Split(':').Last(); // Values are returned with preffixed namespace name, pick only the attribute name
+                if ("name".Equals(attributeName))
+                {
+                    // Check if the activity name is UnityPlayerActivity or UnityPlayerGameActivity
+                    bool isUnityActivity =
+                        k_unityActivityName.Equals(attrib.Value)
+                        || k_unityGameActivityName.Equals(attrib.Value);
+                    Assert.IsTrue(isUnityActivity, "Activity name is not UnityPlayerActivity or UnityPlayerGameActivity");
+                }
+                else if (!expectedElementAttrributes.Contains(new KeyValuePair<string, string>(attributeName, attrib.Value)))
+                {
+                    Assert.Fail($"Unexpected attribute \"{attrib.Name}\" " +
+                        $"with value \"{attrib.Value}\" found in element {node.Name}");
+                }
+            }
+        }
     }
 
     [Test]
@@ -372,6 +513,213 @@ public class AndroidManifestTests
             "Elements exist in the Manifest when expecting 0");
     }
 
+    [Test]
+    public void AndroidManifestProcessor_CheckThatActivityElementHasExportedAttributeWithIntents()
+    {
+        var processor = CreateProcessor();
+
+        var newElementAttributes = new Dictionary<string, string>()
+        {
+            { "name", "com.oculus.intent.category.VR" }
+        };
+        var requirementPrvoider = new List<IAndroidManifestRequirementProvider>()
+        {
+            new MockManifestRequirementProvider(new ManifestRequirement
+            {
+                SupportedXRLoaders = new HashSet<Type>
+                {
+                    typeof(object) // Dummy object representing an inactive loader
+                },
+                NewElements = new List<ManifestElement>()
+                {
+                    new ManifestElement()
+                    {
+                        ElementPath = k_categoryPath,
+                        Attributes = newElementAttributes
+                    }
+                }
+            })
+        };
+
+        processor.ProcessManifestRequirements(requirementPrvoider);
+
+        var xrLibManifest = GetXrLibraryManifest();
+        var activityNodes = xrLibManifest.SelectNodes(string.Join("/", k_activityPath));
+
+        Assert.AreEqual(1, activityNodes.Count, "Expected 1 activity node in the manifest");
+
+        bool foundExportedAttribute = false;
+        foreach (XmlElement activityNode in activityNodes)
+        {
+            var attributeValue = activityNode.GetAttribute("exported", k_androidXmlNamespace);
+            if ("true".Equals(attributeValue))
+            {
+                foundExportedAttribute = true;
+                break;
+            }
+        }
+
+        Assert.IsFalse(foundExportedAttribute, "exported attribute shouldn't be present");
+    }
+
+    [Test]
+    public void AndroidManifestProcessor_CheckThatActivityElementDoesntHaveExportedAttributeWithoutIntents()
+    {
+        var processor = CreateProcessor();
+
+        processor.ProcessManifestRequirements(new List<IAndroidManifestRequirementProvider>());
+
+        var xrLibManifest = GetXrLibraryManifest();
+        var activityNodes = xrLibManifest.SelectNodes(string.Join("/", k_activityPath));
+
+        Assert.AreEqual(1, activityNodes.Count, "Expected 1 activity node in the manifest");
+
+        bool foundExportedAttribute = false;
+        foreach (XmlElement activityNode in activityNodes)
+        {
+            var attributeValue = activityNode.GetAttribute("exported", k_androidXmlNamespace);
+            if ("true".Equals(attributeValue))
+            {
+                foundExportedAttribute = true;
+                break;
+            }
+        }
+
+        Assert.IsFalse(foundExportedAttribute, "exported attribute shouldn't be present");
+    }
+
+    [Test]
+    public void AndroidManifestProcessor_CheckThatGameActivityCanBeCreated()
+    {
+        IgnoreIfGameActivityIsNotSupported();
+
+        var processor = CreateProcessor();
+        processor.UseActivityAppEntry = true;
+        processor.UseGameActivityAppEntry = false;
+
+        processor.ProcessManifestRequirements(new List<IAndroidManifestRequirementProvider>());
+
+        var xrLibManifest = GetXrLibraryManifest();
+        var activityNodes = xrLibManifest.SelectNodes(string.Join("/", k_activityPath));
+
+        Assert.AreEqual(1, activityNodes.Count, "Expected 1 activity node in the manifest");
+
+        bool foundUnityActivity = false;
+        foreach (XmlElement activityNode in activityNodes)
+        {
+            var attributeValue = activityNode.GetAttribute("name", k_androidXmlNamespace);
+            if (k_unityActivityName.Equals(attributeValue))
+            {
+                foundUnityActivity = true;
+                break;
+            }
+        }
+
+        Assert.IsTrue(foundUnityActivity, "UnityPlayerActivity not found in the manifest");
+    }
+
+    [Test]
+    public void AndroidManifestProcessor_CheckThatNormalActivityAndGameActivityCanBeCreated()
+    {
+        IgnoreIfGameActivityIsNotSupported();
+
+        var processor = CreateProcessor();
+        processor.UseActivityAppEntry = true;
+        processor.UseGameActivityAppEntry = true;
+
+        processor.ProcessManifestRequirements(new List<IAndroidManifestRequirementProvider>());
+
+        var xrLibManifest = GetXrLibraryManifest();
+        var activityNodes = xrLibManifest.SelectNodes(string.Join("/", k_activityPath));
+
+        Assert.AreEqual(2, activityNodes.Count, "Expected 2 activity nodes in the manifest");
+
+        bool foundUnityActivity = false;
+        bool foundUnityGameActivity = false;
+        foreach (XmlElement activityNode in activityNodes)
+        {
+            var attributeValue = activityNode.GetAttribute("name", k_androidXmlNamespace);
+            switch (attributeValue)
+            {
+                case k_unityActivityName:
+                    foundUnityActivity = true;
+                    break;
+                case k_unityGameActivityName:
+                    foundUnityGameActivity = true;
+                    break;
+            }
+        }
+
+        Assert.IsTrue(foundUnityActivity, "UnityPlayerActivity not found in the manifest");
+        Assert.IsTrue(foundUnityGameActivity, "UnityPlayerGameActivity not found in the manifest");
+    }
+
+    [Test]
+    public void AndroidManifestProcessor_NewCategoryElementsAreAddedAlongExistingCategoryElements()
+    {
+        var categoryElementName = "category";
+        var processor = CreateProcessor();
+        var categoryPath = new List<string>(k_categoryPath);
+        categoryPath.Append(categoryElementName);
+
+        // Existing Intent-Filter Category element
+        var unityLibManifest = GetUnityLibraryManifest();
+        var categoryAttributes = new Dictionary<string, string>()
+        {
+            { "name", "android.intent.category.LAUNCHER" }
+        };
+        unityLibManifest.CreateNewElement(categoryPath, categoryAttributes);
+        unityLibManifest.Save();
+
+        // New Intent-Filter Category element
+        var newElementAttributes = new Dictionary<string, string>()
+        {
+            { "name", "com.oculus.intent.category.VR" }
+        };
+        var requirementPrvoider = new List<IAndroidManifestRequirementProvider>()
+        {
+            new MockManifestRequirementProvider(new ManifestRequirement
+            {
+                SupportedXRLoaders = new HashSet<Type>
+                {
+                    supportedLoaderType
+                },
+                NewElements = new List<ManifestElement>()
+                {
+                    new ManifestElement()
+                    {
+                        ElementPath = k_categoryPath,
+                        Attributes = newElementAttributes
+                    }
+                }
+            })
+        };
+
+        processor.ProcessManifestRequirements(requirementPrvoider);
+
+        // Reload the manifest
+        unityLibManifest = GetUnityLibraryManifest();
+        var nodes = unityLibManifest.SelectNodes(string.Join("/", k_categoryPath));
+        Assert.AreEqual(
+            2,
+            nodes.Count,
+            "Additional elements exist in the Manifest when expecting 2");
+
+        bool existingCategoryFound = false;
+        bool newCategoryFound = false;
+        foreach (XmlElement node in nodes)
+        {
+            if (categoryElementName.Equals(node.Name))
+            {
+                var categoryAttribValue = node.Attributes.GetNamedItem("name", k_androidXmlNamespace).Value;
+                existingCategoryFound |= "android.intent.category.LAUNCHER".Equals(categoryAttribValue);
+                newCategoryFound |= "com.oculus.intent.category.VR".Equals(categoryAttribValue);
+            }
+        }
+        Assert.IsTrue(existingCategoryFound, "Existing category element not found");
+        Assert.IsTrue(newCategoryFound, "New category element not found");
+    }
+
 #if UNITY_2021_1_OR_NEWER
     [Test]
     public void AndroidManifestProcessor_AddNewIntentsOnlyInUnityLibraryManifest()
@@ -379,8 +727,6 @@ public class AndroidManifestTests
         var processor = CreateProcessor();
 
         // Initialize data
-        var newElementPath =
-            new List<string> { "manifest", "application", "activity", "intent-filter", "category" };
         var newElementAttributes = new Dictionary<string, string>()
         {
             { "name", "com.oculus.intent.category.VR" }
@@ -397,7 +743,7 @@ public class AndroidManifestTests
                 {
                     new ManifestElement()
                     {
-                        ElementPath = newElementPath,
+                        ElementPath = k_categoryPath,
                         Attributes = newElementAttributes
                     }
                 }
@@ -407,7 +753,7 @@ public class AndroidManifestTests
         // Execute
         processor.ProcessManifestRequirements(providers);
 
-        var elementPath = string.Join("/", newElementPath);
+        var elementPath = string.Join("/", k_categoryPath);
 
         // Validate that the intent is created in Unity library manifest
         var unityLibManifest = GetUnityLibraryManifest();
@@ -432,8 +778,6 @@ public class AndroidManifestTests
         var processor = CreateProcessor();
 
         // Initialize data
-        var newElementPath =
-            new List<string> { "manifest", "application", "activity", "intent-filter", "category" };
         var newElementAttributes = new Dictionary<string, string>()
         {
             { "name", "com.oculus.intent.category.VR" }
@@ -450,7 +794,7 @@ public class AndroidManifestTests
                 {
                     new ManifestElement()
                     {
-                        ElementPath = newElementPath,
+                        ElementPath = k_categoryPath,
                         Attributes = newElementAttributes
                     }
                 }
@@ -459,13 +803,13 @@ public class AndroidManifestTests
 
         // Prepare test document
         var appManifest = GetUnityLibraryManifest();
-        appManifest.CreateNewElement(newElementPath, newElementAttributes);
+        appManifest.CreateNewElement(k_categoryPath, newElementAttributes);
         appManifest.Save();
 
         // Execute
         processor.ProcessManifestRequirements(providers);
 
-        var elementPath = string.Join("/", newElementPath);
+        var elementPath = string.Join("/", k_categoryPath);
 
         // Validate that only one intent of the same kind is in the manifest
         var unityLibManifest = GetUnityLibraryManifest();
@@ -483,8 +827,6 @@ public class AndroidManifestTests
         var processor = CreateProcessor();
 
         // Initialize data
-        var newElementPath =
-            new List<string> { "manifest", "application", "activity", "intent-filter", "category" };
         var newElementAttributes = new Dictionary<string, string>()
         {
             { "name", "com.oculus.intent.category.VR" }
@@ -501,12 +843,12 @@ public class AndroidManifestTests
                 {
                     new ManifestElement()
                     {
-                        ElementPath = newElementPath,
+                        ElementPath = k_categoryPath,
                         Attributes = newElementAttributes
                     },
                     new ManifestElement()
                     {
-                        ElementPath = newElementPath,
+                        ElementPath = k_categoryPath,
                         Attributes = newElementAttributes
                     }
                 }
@@ -516,7 +858,7 @@ public class AndroidManifestTests
         // Execute
         processor.ProcessManifestRequirements(providers);
 
-        var elementPath = string.Join("/", newElementPath);
+        var elementPath = string.Join("/", k_categoryPath);
 
         // Validate that only one intent of the same kind is in the manifest
         var unityLibManifest = GetUnityLibraryManifest();
@@ -560,7 +902,7 @@ public class AndroidManifestTests
     {
         var manifestDocument = new AndroidManifestDocument();
         var manifestNode = manifestDocument.CreateElement("manifest");
-        manifestNode.SetAttribute("xmlns:android", "http://schemas.android.com/apk/res/android");
+        manifestNode.SetAttribute("xmlns:android", k_androidXmlNamespace);
         manifestDocument.AppendChild(manifestNode);
         var applicationNode = manifestDocument.CreateElement("application");
         manifestNode.AppendChild(applicationNode);
@@ -588,6 +930,13 @@ public class AndroidManifestTests
         return new List<Dictionary<TKey, TValue>> { dict1, dict2 }
         .SelectMany(dict => dict)
         .ToDictionary(pair => pair.Key, pair => pair.Value);
+    }
+
+    private void IgnoreIfGameActivityIsNotSupported()
+    {
+#if !UNITY_2023_1_OR_NEWER
+        Assert.Ignore("Ignoring test as GameActivity is not supported in Unity versions before 2023.1");
+#endif
     }
 
     private class MockManifestRequirementProvider : IAndroidManifestRequirementProvider